% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRiskSummaryPlot.R
\name{genomicRiskSummaryPlot}
\alias{genomicRiskSummaryPlot}
\title{Function to generate the genomic risk summary plot, as seen in the README.}
\usage{
genomicRiskSummaryPlot(data, column)
}
\arguments{
\item{data}{A data.frame of the variables to use in the plots. The column that is used for the line plot must
have one additional value than the other columns. The names of the columns in the data.frame will be
used for the names of the columns in the plot.}

\item{column}{The name or index of the column to be used for the line plot.}
}
\description{
A function to generate the genomic risk summary plot, which is a series of text plots accompanied
by a line plot displaying the values from one of the columns, usually post-test probability.
}
\examples{
require(ggplot2); require(grid); require(gridExtra); require(scales);

MI_df <- read.csv(system.file("extdata", "MI_genomic_risk.txt", package = "ClinPlots"), sep='\\t')
genomicRiskSummaryPlot(MI_df, "pt_probability")
#running genomicRiskSummaryPlot(MI_df, 7) will produce same result
}

