% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disorderCohortRiskPlot.R
\name{disorderCohortRiskPlot}
\alias{disorderCohortRiskPlot}
\title{Function to generate the disorder risk cohort plot of post-test probabilities, as seen in the README.}
\usage{
disorderCohortRiskPlot(data)
}
\arguments{
\item{data}{A data.frame that contains the columns disorder, n, min, max, avg, and increased. The disorder column
contains the name of the disorder (character), the n column contains the number of individuals in the disorder
cohort (integer), the min column contains the minimum post-test probability (numeric), the max columns
contains the maximum post-test probability (numeric), the avg column contains the average post-test
probability (numeric), and increased is a column that states if the post-test probability increased (logical).}
}
\description{
A function to generate the disorder risk cohort plot, which summarizes the the post-test
             probabilities of each disorder cohort. The disorder risk cohort plot is comprised of a line plot
             that displays the distribution of post-test probabilities for each disorder, along with the
             direction of change in probability. Each line in the plot is aligned with the name of the disorder,
             as well as the number of individuals in the disorder cohort.
}
\examples{
clinical_risk <- read.csv(system.file("extdata", "clinical_risk.txt", package = "ClinPlots"), sep='\\t')
disorderCohortRiskPlot(clinical_risk)
}

